third_party/WebKit/LayoutTests/fast/shapes/shape-outside-floats/shape-outside-floats-ellipse-margin-left-expected.txt

The left edges of the three black squares should follow the outer ellipse boundary and each square should appear on a subsequent line.

X
X
X
PASS elementRect('s1').top is 0
FAIL elementRect('s1').left should be 347. Was 347.09375.
PASS elementRect('s2').top is 20
FAIL elementRect('s2').left should be 390. Was 389.984375.
PASS elementRect('s3').top is 40
FAIL elementRect('s3').left should be 417. Was 417.1875.
PASS successfullyParsed is true

TEST COMPLETE
